fix: remove modal
