+++ packages/components/package.json
@@ -15,7 +15,6 @@
     "react-native-draggable-flatlist": "4.0.0",
     "react-native-keyboard-aware-scroll-view": "^0.9.5",
     "react-native-markdown-display": "^7.0.0-alpha.2",
-    "react-native-modal": "^13.0.1",
     "react-native-modalize": "^2.1.1",
     "react-native-progress": "^5.0.0",
     "react-native-root-siblings": "^4.1.1",
