+++ packages/components/src/Select/Container/Mobile.tsx
@@ -1,8 +1,9 @@
-import { isValidElement } from 'react';
+import { isValidElement, useEffect, useMemo, useRef } from 'react';
 
-import Modal from 'react-native-modal';
+import { Modalize } from 'react-native-modalize';
 
-import { useSafeAreaInsets } from '@onekeyhq/components';
+import { useSafeAreaInsets, useThemeValue } from '@onekeyhq/components';
+import { PortalEntry } from '@onekeyhq/kit/src/views/Overlay/RootPortal';
 
 import Box from '../../Box';
 import Button from '../../Button';
@@ -31,58 +32,107 @@ function Mobile<T>({
   activatable,
 }: ChildProps<T>) {
   const { bottom } = useSafeAreaInsets();
-  return (
-    <Modal
-      propagateSwipe
-      hideModalContentWhileAnimating
-      useNativeDriver
-      useNativeDriverForBackdrop
-      isVisible={!!visible}
-      onModalHide={onModalHide}
-      swipeDirection={['down']}
-      onSwipeComplete={toggleVisible}
-      onBackdropPress={toggleVisible}
-      animationInTiming={150}
-      animationOutTiming={150}
-      animationIn="slideInUp"
-      animationOut="slideOutDown"
-      style={{
-        justifyContent: 'flex-end',
-        margin: 0,
+  const modalizeRef = useRef<Modalize>(null);
+  const [defaultBgColor, handleBgColor] = useThemeValue([
+    'background-default',
+    'icon-subdued',
+  ]);
+
+  useEffect(() => {
+    if (visible) {
+      setTimeout(() => modalizeRef.current?.open(), 10);
+    } else {
+      onModalHide?.();
+      modalizeRef.current?.close();
+    }
+  }, [visible, onModalHide]);
+
+  const headerComponent = useMemo(
+    () => (
+      <Box
+        py="1"
+        px="2"
+        display="flex"
+        flexDirection="row"
+        justifyContent="space-between"
+        alignItems="center"
+        bg="transparent"
+      >
+        {/* placeholder */}
+        <Box width="12" />
+        <Box flex="1" flexDirection="row" justifyContent="center">
+          <Typography.Heading numberOfLines={1}>{title}</Typography.Heading>
+        </Box>
+        <IconButton
+          name="XMarkOutline"
+          type="plain"
+          size="xl"
+          onPress={toggleVisible}
+          circle
+        />
+      </Box>
+    ),
+    [title, toggleVisible],
+  );
+
+  const footerComponent = useMemo(
+    () =>
+      isValidElement(footer) || footer === null ? (
+        footer
+      ) : (
+        <Box pb={`${bottom}px`} bg="surface-subdued">
+          <Box px="4" py="2" borderTopWidth={1} borderTopColor="border-subdued">
+            <Button
+              size="xl"
+              type="plain"
+              leftIconName={footerIcon}
+              onPress={onPressFooter}
+            >
+              {footerText}
+            </Button>
+          </Box>
+        </Box>
+      ),
+    [bottom, footer, footerIcon, footerText, onPressFooter],
+  );
+
+  const content = (
+    <Modalize
+      adjustToContentHeight
+      ref={modalizeRef}
+      withHandle={false}
+      onClose={onModalHide}
+      modalStyle={{
+        backgroundColor: defaultBgColor,
+        borderTopLeftRadius: 24,
+        borderTopRightRadius: 24,
       }}
+      handlePosition="inside"
+      handleStyle={{
+        // default styles start
+        alignSelf: 'center',
+        top: 8,
+        width: 45,
+        height: 5,
+        borderRadius: 5,
+        // default styles end
+
+        // custom styles
+        backgroundColor: handleBgColor,
+      }}
+      HeaderComponent={headerComponent}
+      FooterComponent={footerComponent}
     >
       <Box
-        maxHeight="70%"
+        // maxHeight="70%"
         minHeight="180px"
         minW="full"
-        bg="surface-subdued"
+        // bg="surface-subdued"
         borderTopRadius="24px"
-        pb={`${footer === null ? bottom : 0}px`}
+        pb="0px"
+        // pb={`${footer === null ? bottom : 0}px`}
         {...dropdownProps}
       >
-        <Box
-          py="1"
-          px="2"
-          display="flex"
-          flexDirection="row"
-          justifyContent="space-between"
-          alignItems="center"
-          borderBottomColor="border-subdued"
-          borderBottomWidth={title ? 1 : undefined}
-        >
-          {/* placeholder */}
-          <Box width="12" />
-          <Box flex="1" flexDirection="row" justifyContent="center">
-            <Typography.Heading numberOfLines={1}>{title}</Typography.Heading>
-          </Box>
-          <IconButton
-            name="XMarkOutline"
-            type="plain"
-            size="xl"
-            onPress={toggleVisible}
-            circle
-          />
-        </Box>
         <ScrollView _contentContainerStyle={{ padding: 2, paddingBottom: '4' }}>
           <RenderOptions
             options={options}
@@ -92,30 +142,11 @@ function Mobile<T>({
             activatable={activatable}
           />
         </ScrollView>
-        {isValidElement(footer) || footer === null ? (
-          footer
-        ) : (
-          <Box pb={`${bottom}px`}>
-            <Box
-              px="4"
-              py="2"
-              borderTopWidth={1}
-              borderTopColor="border-subdued"
-            >
-              <Button
-                size="xl"
-                type="plain"
-                leftIconName={footerIcon}
-                onPress={onPressFooter}
-              >
-                {footerText}
-              </Button>
-            </Box>
-          </Box>
-        )}
       </Box>
-    </Modal>
+    </Modalize>
   );
+
+  return <PortalEntry target="Root-FullWindowOverlay">{content}</PortalEntry>;
 }
 
 export default Mobile;
