+++ packages/kit/src/provider/KitProvider.tsx
@@ -1,22 +1,16 @@
 import type { FC } from 'react';
-import { useMemo } from 'react';
 
 import axios from 'axios';
-import { StyleSheet, View } from 'react-native';
 import { GestureHandlerRootView } from 'react-native-gesture-handler';
-import { RootSiblingParent } from 'react-native-root-siblings';
-import { FullWindowOverlay } from 'react-native-screens';
 import { Provider as ReduxProvider } from 'react-redux';
 import { SWRConfig } from 'swr';
 
-import CustomToast from '@onekeyhq/components/src/Toast/Custom';
 import type { NotificationExtra } from '@onekeyhq/engine/src/managers/notification';
 import { ErrorBoundary } from '@onekeyhq/kit/src/components/ErrorBoundary';
 import store from '@onekeyhq/kit/src/store';
 import platformEnv from '@onekeyhq/shared/src/platformEnv';
 
 import { createLazyComponent } from '../utils/createLazyComponent';
-import { PortalExit } from '../views/Overlay/RootPortal';
 
 import AppLoading from './AppLoading';
 import NavigationProvider from './NavigationProvider';
@@ -59,16 +53,6 @@ const KitProvider: FC<LaunchProps> = (propsRaw) => {
     name: 'KitProvider render',
     payload: props,
   });
-  const globalPortalViews = useMemo(
-    () => (
-      <View pointerEvents="box-none" style={StyleSheet.absoluteFill}>
-        <ChainWebEmbed />
-        <CustomToast bottomOffset={60} />
-        <PortalExit name="Root-FullWindowOverlay" />
-      </View>
-    ),
-    [],
-  );
   return (
     <SWRConfig value={swrConfig}>
       <ReduxProvider store={store}>
@@ -77,19 +61,9 @@ const KitProvider: FC<LaunchProps> = (propsRaw) => {
             <AppLoading>
               <ErrorBoundary>
                 <NotificationProvider launchNotification={launchNotification} />
-                <RootSiblingParent>
-                  <NavigationProvider />
-                  <WhenAppActive />
-                  {platformEnv.isNativeIOS ? (
-                    // FullWindowOverlay can render above native views
-                    // but can not work with modal
-                    // https://github.com/software-mansion/react-native-screens/issues/1149
-                    // so now only used for toast
-                    <FullWindowOverlay>{globalPortalViews}</FullWindowOverlay>
-                  ) : (
-                    globalPortalViews
-                  )}
-                </RootSiblingParent>
+                <ChainWebEmbed />
+                <NavigationProvider />
+                <WhenAppActive />
               </ErrorBoundary>
             </AppLoading>
           </ThemeProvider>
