+++ packages/kit/src/provider/NavigationProvider.tsx
@@ -2,9 +2,12 @@ import { createRef, memo, useEffect, useMemo, useRef } from 'react';
 
 import { useFlipper } from '@react-navigation/devtools';
 import { DefaultTheme, NavigationContainer } from '@react-navigation/native';
+import { StyleSheet, View } from 'react-native';
 import { RootSiblingParent } from 'react-native-root-siblings';
+import { FullWindowOverlay } from 'react-native-screens';
 
 import { useIsVerticalLayout, useThemeValue } from '@onekeyhq/components';
+import CustomToast from '@onekeyhq/components/src/Toast/Custom';
 import { useSettings } from '@onekeyhq/kit/src/hooks/redux';
 import { RootStackNavigator } from '@onekeyhq/kit/src/routes';
 import type { RootRoutesParams } from '@onekeyhq/kit/src/routes/types';
@@ -16,6 +19,7 @@ import platformEnv from '@onekeyhq/shared/src/platformEnv';
 import { useShortcuts } from '../hooks/useShortcuts';
 import '../routes/deepLink';
 import buildLinking from '../routes/linking';
+import { PortalExit } from '../views/Overlay/RootPortal';
 
 import RedirectProvider from './RedirectProvider';
 
@@ -96,6 +100,16 @@ const NavigationApp = () => {
 
   useShortcuts();
 
+  const globalPortalViews = useMemo(
+    () => (
+      <View pointerEvents="box-none" style={StyleSheet.absoluteFill}>
+        <CustomToast bottomOffset={60} />
+        <PortalExit name="Root-FullWindowOverlay" />
+      </View>
+    ),
+    [],
+  );
+
   return (
     <NavigationContainer
       documentTitle={{
@@ -132,6 +146,15 @@ const NavigationApp = () => {
         <RedirectProvider>
           <RootStackNavigator />
         </RedirectProvider>
+        {platformEnv.isNativeIOS ? (
+          // FullWindowOverlay can render above native views
+          // but can not work with modal
+          // https://github.com/software-mansion/react-native-screens/issues/1149
+          // so now only used for toast
+          <FullWindowOverlay>{globalPortalViews}</FullWindowOverlay>
+        ) : (
+          globalPortalViews
+        )}
       </RootSiblingParent>
     </NavigationContainer>
   );
