+++ packages/kit/src/routes/Root/Modal/index.tsx
@@ -1,9 +1,6 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import { memo, useMemo } from 'react';
 
-// import { createStackNavigator } from '@react-navigation/stack';
-import { RootSiblingParent } from 'react-native-root-siblings';
-
 import { useIsVerticalLayout } from '@onekeyhq/components';
 
 import { createLazyComponent } from '../../../utils/createLazyComponent';
@@ -251,17 +248,15 @@ const ModalStackNavigator = () => {
     [isVerticalLayout],
   );
   return (
-    <RootSiblingParent>
-      <ModalStack.Navigator screenOptions={screenOptions}>
-        {modalStackScreenList.map((modal) => (
-          <ModalStack.Screen
-            key={modal.name}
-            name={modal.name}
-            component={modal.component}
-          />
-        ))}
-      </ModalStack.Navigator>
-    </RootSiblingParent>
+    <ModalStack.Navigator screenOptions={screenOptions}>
+      {modalStackScreenList.map((modal) => (
+        <ModalStack.Screen
+          key={modal.name}
+          name={modal.name}
+          component={modal.component}
+        />
+      ))}
+    </ModalStack.Navigator>
   );
 };
