+++ packages/kit/src/views/ManageTokens/Listing.tsx
@@ -561,48 +561,45 @@ export const Listing: FC = () => {
       if (!token) {
         return;
       }
-      showOverlay(
-        (closeOverlay) => (
-          <Dialog
-            visible
-            onClose={closeOverlay}
-            footerButtonProps={{
-              primaryActionTranslationId: 'action__delete',
-              primaryActionProps: {
-                type: 'destructive',
-                onPromise: async () => {
-                  if (accountId && token) {
-                    await backgroundApiProxy.engine.removeTokenFromAccount(
-                      accountId,
-                      token.id,
-                    );
-                  }
-                  await backgroundApiProxy.serviceToken.fetchAccountTokens({
+      showOverlay((closeOverlay) => (
+        <Dialog
+          visible
+          onClose={closeOverlay}
+          footerButtonProps={{
+            primaryActionTranslationId: 'action__delete',
+            primaryActionProps: {
+              type: 'destructive',
+              onPromise: async () => {
+                if (accountId && token) {
+                  await backgroundApiProxy.engine.removeTokenFromAccount(
                     accountId,
-                    networkId,
-                  });
-                  closeOverlay();
-                },
+                    token.id,
+                  );
+                }
+                await backgroundApiProxy.serviceToken.fetchAccountTokens({
+                  accountId,
+                  networkId,
+                });
+                closeOverlay();
               },
-            }}
-            contentProps={{
-              iconType: 'danger',
-              title: intl.formatMessage({
-                id: 'modal__delete_this_token',
-                defaultMessage: 'Delete this token?',
-              }),
-              content: intl.formatMessage(
-                {
-                  id: 'modal__delete_this_token_desc',
-                  defaultMessage: '{token} will be removed from my tokens',
-                },
-                { token: token.name },
-              ),
-            }}
-          />
-        ),
-        true,
-      );
+            },
+          }}
+          contentProps={{
+            iconType: 'danger',
+            title: intl.formatMessage({
+              id: 'modal__delete_this_token',
+              defaultMessage: 'Delete this token?',
+            }),
+            content: intl.formatMessage(
+              {
+                id: 'modal__delete_this_token_desc',
+                defaultMessage: '{token} will be removed from my tokens',
+              },
+              { token: token.name },
+            ),
+          }}
+        />
+      ));
     },
     [accountId, intl, networkId],
   );
